:art: Move all test fixtures into `fixtures` directory
