+++ .gitattributes
@@ -1,5 +1,3 @@
 *.ts linguist-language=TypeScript
-test/npm/** linguist-detectable=false
-test/yarn/** linguist-detectable=false
-test/pnpm/** linguist-detectable=false
 *.config.js linguist-detectable=false
+test/fixtures/** linguist-detectable=false
