+++ test/fixtures/npm/deps/index.js
