+++ test/fixtures/npm/deps/package.json
