+++ test/fixtures/npm/dev-deps/index.js
