+++ test/fixtures/npm/dev-deps/package.json
