+++ test/fixtures/pnpm/deps/index.js
