+++ test/fixtures/pnpm/deps/package.json
