+++ test/fixtures/pnpm/dev-deps/index.js
