+++ test/fixtures/pnpm/dev-deps/package.json
