+++ test/fixtures/yarn/deps/index.js
